feat: add BETTER portfolio analysis
style: sate flake8

The taste
Of rain
—Why kneel?
