+++ seed/analysis_pipelines/better/__init__.py
@@ -1,0 +1,7 @@
+from seed.analysis_pipelines.better.pipeline import (  # noqa: F401
+    BETTERPipeline,
+    _prepare_all_properties,
+    _finish_preparation,
+    _start_analysis,
+    _finish_analysis,
+)
