+++ seed/analysis_pipelines/better/buildingsync.py
@@ -1,0 +1,201 @@
+from lxml import etree
+from lxml.builder import ElementMaker
+
+from seed.analysis_pipelines.pipeline import AnalysisPipelineException
+from seed.building_sync.mappings import BUILDINGSYNC_URI, NAMESPACES
+
+
+# PREMISES_ID_NAME is the name of the custom ID used within a BuildingSync document
+# to link it to SEED's AnalysisPropertyViews
+PREMISES_ID_NAME = 'seed_analysis_property_view_id'
+
+# BETTER and ESPM use different names for property types than BEDES and BSync
+BETTER_TO_BSYNC_PROPERTY_TYPE = {
+    'Office': 'Office',
+    'Hotel': 'Lodging',
+    'K-12 School': 'Education',
+    'Hospital (General Medical & Surgical)': 'Health care-Inpatient hospital',
+    'Bank Branch': 'Bank',
+    'Courthouse': 'Courthouse',
+    'Data Center': 'Data Center',
+    'Distribution Center': 'Distribution Center',
+    'Financial Office': 'Office-Financial',
+    'Multifamily Housing': 'Multifamily',
+    'Non-Refrigerated Warehouse': 'Warehouse unrefrigerated',
+    'Refrigerated Warehouse': 'Warehouse refrigerated',
+    'Retail Store': 'Retail',
+    'Senior Care Community': 'Health care-Skilled nursing facility',
+    'Supermarket/Grocery Store': 'Food sales-Grocery store',
+    'Other': 'Other'
+}
+
+
+def _build_better_input(analysis_property_view, meters):
+    """Constructs a BuildingSync document to be used as input for a BETTER analysis.
+    The function returns a tuple, the first value being the XML document as a byte
+    string. The second value is a list of error messages.
+
+    :param analysis_property_view: AnalysisPropertyView
+    :param meter: Meter
+    :returns: tuple(bytes, list[str])
+    """
+    # TODO Refine ID assignment
+    errors = []
+    property_state = analysis_property_view.property_state
+
+    if property_state.property_name is None:
+        errors.append("BETTER analysis requires the property's name.")
+    if property_state.city is None:
+        errors.append("BETTER analysis requires the property's city.")
+    if property_state.state is None:
+        errors.append("BETTER analysis requires the property's state.")
+    if property_state.gross_floor_area is None:
+        errors.append("BETTER analysis requires the property's gross floor area.")
+    if property_state.property_type is None:
+        errors.append("BETTER analysis requires the property's type (office, retail, etc).")
+    if property_state.property_type not in BETTER_TO_BSYNC_PROPERTY_TYPE:
+        errors.append(
+            f"BETTER analysis requires the property's type must be one of the following: {', '.join(BETTER_TO_BSYNC_PROPERTY_TYPE.keys())}")
+    for meter in meters:
+        for meter_reading in meter.meter_readings.all():
+            if meter_reading.reading is None:
+                errors.append(f'{meter}: MeterReading starting at {meter_reading.start_time} has no reading value.')
+    if errors:
+        return None, errors
+
+    # clean inputs
+    # BETTER will default if eGRIDRegion is not explicitly set
+    try:
+        eGRIDRegion = property_state.extra_data['eGRIDRegion']
+    except KeyError:
+        eGRIDRegion = ""
+
+    property_type = BETTER_TO_BSYNC_PROPERTY_TYPE[property_state.property_type]
+
+    gross_floor_area = str(int(property_state.gross_floor_area.magnitude))
+
+    XSI_URI = 'http://www.w3.org/2001/XMLSchema-instance'
+    nsmap = {
+        'xsi': XSI_URI,
+    }
+    nsmap.update(NAMESPACES)
+    E = ElementMaker(
+        namespace=BUILDINGSYNC_URI,
+        nsmap=nsmap
+    )
+
+    doc = (
+        E.BuildingSync(
+            {
+                etree.QName(XSI_URI,
+                            'schemaLocation'): 'http://buildingsync.net/schemas/bedes-auc/2019 https://raw.github.com/BuildingSync/schema/v2.2.0/BuildingSync.xsd',
+                'version': '2.2.0'
+            },
+            E.Facilities(
+                E.Facility(
+                    {'ID': 'Facility-1'},
+                    E.Sites(
+                        E.Site(
+                            {'ID': 'Site-1'},
+                            E.Buildings(
+                                E.Building(
+                                    {'ID': 'Building-1'},
+                                    E.PremisesName(property_state.property_name),
+                                    E.PremisesIdentifiers(
+                                        E.PremisesIdentifier(
+                                            E.IdentifierLabel('Custom'),
+                                            E.IdentifierCustomName(PREMISES_ID_NAME),
+                                            E.IdentifierValue(str(analysis_property_view.id)),
+                                        )
+                                    ),
+                                    E.Address(
+                                        E.StreetAddressDetail(
+                                            E.Simplified(
+                                                E.StreetAddress(str(property_state.address_line_1))
+                                            )
+                                        ),
+                                        E.City(property_state.city),
+                                        E.State(property_state.state),
+                                        E.PostalCode(property_state.postal_code)
+                                    ),
+                                    E.eGRIDRegionCode(eGRIDRegion),
+                                    E.Longitude(str(analysis_property_view.property_state.longitude)),
+                                    E.Latitude(str(analysis_property_view.property_state.latitude)),
+                                    E.OccupancyClassification(property_type),
+                                    E.FloorAreas(
+                                        E.FloorArea(
+                                            E.FloorAreaType("Gross"),
+                                            E.FloorAreaValue(gross_floor_area)
+                                        )
+                                    ),
+                                )
+                            )
+                        )
+                    ),
+                    E.Reports(
+                        E.Report(
+                            {'ID': 'Report-1'},
+                            E.Scenarios(
+                                E.Scenario(
+                                    {'ID': 'Scenario-Measured'},
+                                    E.ScenarioType(
+                                        E.CurrentBuilding(
+                                            E.CalculationMethod(
+                                                E.Measured()
+                                            )
+                                        )
+                                    ),
+                                    E.ResourceUses(
+                                        E.ResourceUse(
+                                            {'ID': 'Resource-' + str(11)},
+                                            E.EnergyResource('Electricity'),
+                                            E.ResourceUnits('kWh'),
+                                            E.EndUse('All end uses')
+                                        ),
+                                        E.ResourceUse(
+                                            {'ID': 'Resource-' + str(19)},
+                                            E.EnergyResource('Natural gas'),
+                                            E.ResourceUnits('MMBtu'),
+                                            E.EndUse('Heating')
+                                        ),
+                                    ),
+                                    E.TimeSeriesData(
+                                        *[
+                                            E.TimeSeries(
+                                                {'ID': f'TimeSeries-{meter.type}-{i}'},
+                                                E.ReadingType('Total'),
+                                                E.StartTimestamp(reading.start_time.isoformat()),
+                                                E.EndTimestamp(reading.end_time.isoformat()),
+                                                E.IntervalFrequency('Month'),
+                                                E.IntervalReading(str(reading.reading)),
+                                                E.ResourceUseID({'IDref': 'Resource-' + str(meter.type)}),
+                                            )
+                                            for meter in meters for i, reading in enumerate(meter.meter_readings.all())
+                                        ]
+                                    ),
+                                    E.LinkedPremises(
+                                        E.Building(
+                                            E.LinkedBuildingID({'IDref': 'Building-1'})
+                                        )
+
+                                    )
+                                )
+                            )
+                        )
+                    )
+                )
+            )
+        )
+    )
+    return etree.tostring(doc, pretty_print=True), []
+
+
+def _parse_analysis_property_view_id(filepath):
+    input_file_tree = etree.parse(filepath)
+    id_xpath = f'//auc:PremisesIdentifier[auc:IdentifierCustomName = "{PREMISES_ID_NAME}"]/auc:IdentifierValue'
+    analysis_property_view_id_elem = input_file_tree.xpath(id_xpath, namespaces=NAMESPACES)
+
+    if len(analysis_property_view_id_elem) != 1:
+        raise AnalysisPipelineException(
+            f'Expected BuildlingSync file to have exactly one "{PREMISES_ID_NAME}" PremisesIdentifier')
+    return int(analysis_property_view_id_elem[0].text)
