+++ seed/static/seed/locales/en_US.json
@@ -943,7 +943,6 @@
     "View my properties": "View my properties",
     "View Project": "View Project",
     "Viewer": "Viewer",
-    "Visit the BETTER website to run a portfolio analysis": "Visit the BETTER website to run a portfolio analysis",
     "Warning": "Warning",
     "We've detected an outdated browser.": "We've detected an outdated browser.",
     "We've e-mailed instructions for resetting your password.": "We've e-mailed instructions for resetting your password.",
