+++ seed/static/seed/locales/fr_CA.json
@@ -943,7 +943,6 @@
     "View my properties": "Voir mes propriétés",
     "View Project": "Voir le projet",
     "Viewer": "Spectateur",
-    "Visit the BETTER website to run a portfolio analysis": "Visitez le site Web de BETTER pour effectuer une analyse de portefeuille",
     "Warning": "Mise en garde",
     "We've detected an outdated browser.": "Nous avons détecté un navigateur obsolète.",
     "We've e-mailed instructions for resetting your password.": "Nous avons envoyé des instructions par e-mail pour réinitialiser votre mot de passe.",
