+++ seed/static/seed/partials/analysis.html
@@ -6,9 +6,6 @@
             <div class="left section_action_container">
                 <h2><i class="fa fa-bar-chart"></i>{$:: 'Analyses' | translate $}: {$:: analysis.name $}</h2>
             </div>
-            <div class="right section_action_container" ng-if="analysis.service == 'BETTER'">
-                <a href="https://better.lbl.gov/" target="_blank">{$:: 'Visit the BETTER website to run a portfolio analysis' | translate $}<i class="fa fa-arrow-right"></i></a>
-            </div>
         </div>
     </div>
     <div class="section_content_container">
