+++ seed/static/seed/partials/inventory_detail_analyses_modal.html
@@ -13,7 +13,7 @@
                 </div>
                 <div class="form-group">
                     <label class="control-label sectionLabel">Type</label>
-                    <select class="form-control" ng-model="new_analysis.service" ng-dropdown required>
+                    <select class="form-control" ng-model="new_analysis.service" ng-dropdown ng-change="initializeAnalysisConfig()" required>
                         <option ng-option value='BSyncr' selected="selected">BSyncr</option>
                         <option ng-option value='BETTER' selected="selected">BETTER</option>
                     </select>
@@ -31,7 +31,10 @@
                     <div class="form-group" style="padding-top: 1em">
                         <div class="form-row">
                             <div class="col-md-6">
-                                <label class="control-label sectionLabel text-muted">Savings Target</label>
+                                <label class="control-label sectionLabel text-muted">
+                                    Savings Target
+                                    <i class="ui-grid-icon-info-circled" uib-tooltip-html="'Default: BETTER will benchmark your building(s) against built-in benchmark statistics derived from a sample of U.S. buildings that match your property types.<br><br>Generate: BETTER will generate benchmark statistics based on the set of buildings being analyzed, so your buildings will be compared against others in your portfolio. This option will provide more accurate statistics if you selected at least 30 buildings for analysis.'"></i>
+                                </label>
                                 <select class="form-control" ng-model="new_analysis.configuration.savings_target" ng-dropdown required>
                                     <option ng-repeat="model in better_savings_targets" selected="selected"
                                             ng-value=model>{$:: model.savings_target $}
@@ -39,7 +42,10 @@
                                 </select>
                             </div>
                             <div class="col-md-6">
-                                <label class="control-label sectionLabel text-muted">Benchmark Data</label>
+                                <label class="control-label sectionLabel text-muted">
+                                    Benchmark Data
+                                    <i class="ui-grid-icon-info-circled" uib-tooltip-html="'Conservative: The savings goal will be one standard deviation worse than the median performance of the benchmarking peer group.<br><br>Nominal: The savings goal will be equal to the median performance of the benchmarking peer group.<br><br>Aggressive: The savings goal will be one half standard deviation better than the median performance of the benchmarking peer group.'"></i>
+                                </label>
                                 <select class="form-control" ng-model="new_analysis.configuration.benchmark_data" ng-dropdown required>
                                     <option ng-repeat="model in better_benchmark_options" selected="selected"
                                             ng-value=model>{$:: model.benchmark_data $}
@@ -48,9 +54,19 @@
                             </div>
                         </div>
                         <div class="form-row">
-                            <div class="col-md-12" style="padding-top: 1em; padding-bottom: 1em">
-                                <label class="control-label sectionLabel text-muted">Minimum Model R Squared</label>
+                            <div class="col-md-6" style="padding-top: 1em; padding-bottom: 1em">
+                                <label class="control-label sectionLabel text-muted">
+                                    Minimum Model R Squared
+                                    <i class="ui-grid-icon-info-circled" uib-tooltip-html="'R² is the proportion of the variance in the dependent variable (energy use) that is predictable from the independent variable (outdoor air temperature).'"></i>
+                                </label>
                                 <input ng-model="new_analysis.configuration.min_r_squared" type="number" class="form-control" step=".01" min="0" max="1" >
+                            </div>
+                            <div class="col-md-6" style="padding-top: 1em; padding-bottom: 1em">
+                                <label class="control-label sectionLabel text-muted">
+                                    Run Portfolio Analysis
+                                    <i class="ui-grid-icon-info-circled" uib-tooltip-html="'Only available if more than one properties were selected for analysis from the inventory list page'"></i>
+                                </label>
+                                <input ng-model="new_analysis.configuration.portfolio_analysis" ng-disabled="inventory_count == 1" type="checkbox" style="width: 34px; height: 34px; display: block; margin: 0;">
                             </div>
                         </div>
                     </div>
